This command is useful for DB cloning
It was hard to find on the internet how to clone DBs. The answer is by using this command + the ".show database schema command", but we didn't write it explicitly... I added it as a "tip". I also added tip about cloning a cluster. This is more an Azure topic (using ARM templates), so I linked the best doc that I could find