Add double backslash for meta-character uses in regex docs
In the regex docs (https://docs.microsoft.com/en-us/azure/data-explorer/kusto/query/re2-library) it says to use only a single backslash to match a meta-character. But this should be changed to double backslash (\\+).
The correct pattern is added in source, but markdown identifies it in the wrong way. So adding four backslashes(////) in the source to make it appear correctly in the docs.