Update output for the query example
The example query returns a table with a single column and a single row containing the count of the filtered rows. The filters applied in the query are for a specific time range and  for a specific state. The refined output contains records only for the specified time range and the specified state.
The count value is 28.