Added the example for the filter for the storage account example
It turns out that the storage account walkthrough that is shown by default did not have the sample and explanation for the filter. I added it here. I would also reconsider making the default view the storage account walkthrough, as I would imagine that most users start from Kusto and not from the storage accounts. 
-------
cc: @orspod; @YaelSchuster ; @shsagir; @vladikbr ; @TzviaGT