Add an example of AddTrustedHosts to the kusto.md
I added an example of the code AddTrustedHosts usage because it is not obvious which namespace should be used and what kind of input should be passed.
I found this namespace in this changelog:
https://github.com/Azure/azure-kusto-dotnet/blob/0348e6dfcf2c9d2918fedf8712587fd8f5f99321/Microsoft.Azure.Kusto.Data.changelog.md
Not sure if this is what is meant here by 'AddTrustedHosts' but I assumed this method. This is the only place I managed to find it.

Side note: is there a certain minimal version of Microsoft.Azure.Kusto.Data that has "Kusto.Data.Common.Impl.KustoTrustedEndpoints" mentioned in the paragraph?  I am using 10.0.2 and my VS does not see it?