Page Order and Language
The order of this page is very confusing. There are many repeating titles such as "Filter by Boolean expression" and so on that provide more information than the previous sections with the same title. This makes the page very long and dense. 
I propose providing an overview/ table of contents for the tutorial at the beginning that depict a logical and cohesive progression. 
For example: 
1. A brief description of what Kusto Query language is. Following that, a seamless continuation into the example used ('StormEvents').
2. Counting.
3. Different operators for picking columns: project/ take. 
4. etc. 
In addition, I would rephrase and organize various sections. The language could be more user friendly in terms of syntax/grammar/ visual presentation.