Update tutorial-archive.md - user session example of join
Need more information on this 'user session example':

It says,  "First,  use `project` to select just the relevant columns before you perform the join." 

Why does the user need to perform the join?

Also, there are two 'project' variables in the query, and both are different. How come? Which one does the user need to refer to in the previous sentence?