Update tutorial-archive.md
Basic summary of changes:
- Fixing grammatical and spelling errors
- Restructuring the document where beneficial to flow and understanding
- Addition of output tables where necessary, removal of irrelevant info
- Addition of introductory / explanatory text where beneficial

Specific commentary on the proposed changes, and some general comments on the document:

There are inconsistencies in how the output tables are shown - the tutorial begins with embedded tables, and then from the “Timechart” sections onwards, the output tables are embedded as images / screenshots. This needs to be shown consistently throughout the document.

I was unable to test the actual code language, as I do not have access to Azure Data or an MS account for logging in. I did however check the query language to the best of my ability within the document, cross referencing other Kusto resources.

Section: “Select a subset of columns: project” section - I would include the output table for the above query, but I am missing this information.

Section: “Aggregate groups of rows: summarize” - I would add an output table.
 
Section “User session example of join” - I cannot include a link to the referenced table as I don’t have access to that information.
 
When I check my proposed changes the embedded images are not working. The link is written out but the image is not embedded in the document. I cannot figure out how to fix this. 
I did try to make adjustments according to Markdown standards, but couldn’t get it exactly right.
This is what I tried, by way of example in the “Display a chart: render” section:
![Image](event-counts-state.png "Event counts") 

There are three places where links to other sections within the document are embedded, but I cannot check if the links are working properly:
In “Assign a result to a variable: *let*” there is a link back to the section “Join data types”
In “Calculate percentiles: *percentile*” there is a link back to the section “Plot a distribution”
The list of topics covered in the tutorial, at the beginning of the document.

7. I felt that if there is a section on joining data from different databases (section: “Combine data from several databases in a query”), there should also be a section on joining data from different tables within the same database. I have added a new section “Combine data from multiple tables: join”. I took some of the information I needed from existing Kusto resources. I felt it was a worthwhile additional section to this tutorial, which is quite an extensive one.  I did not have an output table or chart to include but have indicated where they would appear (shown by {...}).

8. Section “Combine data from several databases in a query”:
I have proposed changes to this section in my PR, but my overall sense is that this section is beyond the scope of this particular tutorial. I would propose that the only text under this section should be: 
“Combining data from several databases typically involves using the `union` operator, which combines the results of two or more queries into a single result set.
For more information about combining data from several databases in a query, see cross-database queries {link to the referenced document}.”