(AzureCXP) fixes MicrosoftDocs/dataexplorer-docs#419684
Updated line no 12.
From: 
Reports the zero-based index of the first occurrence of a specified string within the input string.
To:
Reports the zero-based index of the first occurrence of a specified string within the input string. The `indexof()` function is case-sensitive. To perform a case-insensitive search, consider using `tolower()` or `toupper()` on both inputs.


Updated the missing information.